Move Class to helper concern rename to session handler
